requirements.txt
pandas==1.5.3
requests==2.31.0
streamlit-aggrid==0.3.4.post3